Moved LlmOperations to core.internal package
Also annotated with ApiStatus.Internal.

See gh-1024
Closes gh-1400
