+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/spi/support/LlmRanker.kt
@@ -19,8 +19,8 @@ import com.embabel.agent.api.common.InteractionId
 import com.embabel.agent.api.common.ranking.Ranker
 import com.embabel.agent.api.common.ranking.Ranking
 import com.embabel.agent.api.common.ranking.Rankings
+import com.embabel.agent.core.internal.LlmOperations
 import com.embabel.agent.core.support.LlmInteraction
-import com.embabel.agent.spi.LlmOperations
 import com.embabel.agent.spi.common.RetryProperties
 import com.embabel.common.ai.model.LlmOptions
 import com.embabel.common.ai.model.ModelSelectionCriteria
