+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/test/integration/DummyObjectCreatingLlmOperations.kt
@@ -18,8 +18,8 @@ package com.embabel.agent.test.integration
 import com.embabel.agent.api.event.LlmRequestEvent
 import com.embabel.agent.core.Action
 import com.embabel.agent.core.AgentProcess
+import com.embabel.agent.core.internal.LlmOperations
 import com.embabel.agent.core.support.LlmInteraction
-import com.embabel.agent.spi.LlmOperations
 import com.embabel.chat.Message
 import com.embabel.common.util.DummyInstanceCreator
