+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/test/integration/IntegrationTestUtils.kt
@@ -20,10 +20,10 @@ import com.embabel.agent.api.common.PlatformServices
 import com.embabel.agent.api.event.AgenticEventListener
 import com.embabel.agent.core.*
 import com.embabel.agent.core.expression.LogicalExpressionParser
+import com.embabel.agent.core.internal.LlmOperations
 import com.embabel.agent.core.support.DefaultAgentPlatform
 import com.embabel.agent.core.support.InMemoryBlackboard
 import com.embabel.agent.core.support.SimpleAgentProcess
-import com.embabel.agent.spi.LlmOperations
 import com.embabel.agent.spi.OperationScheduler
 import com.embabel.agent.spi.ToolGroupResolver
 import com.embabel.agent.spi.config.spring.AgentPlatformProperties.ProcessType
