+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/test/integration/ScriptedLlmOperations.kt
@@ -19,8 +19,8 @@ import com.embabel.agent.api.event.LlmRequestEvent
 import com.embabel.agent.api.tool.Tool
 import com.embabel.agent.core.Action
 import com.embabel.agent.core.AgentProcess
+import com.embabel.agent.core.internal.LlmOperations
 import com.embabel.agent.core.support.LlmInteraction
-import com.embabel.agent.spi.LlmOperations
 import com.embabel.chat.Message
 import org.slf4j.LoggerFactory
