+++ embabel-agent-api/src/test/kotlin/com/embabel/agent/api/common/OperationContextPromptRunnerThinkingTest.kt
@@ -21,8 +21,8 @@ import com.embabel.agent.core.AgentPlatform
 import com.embabel.agent.core.AgentProcess
 import com.embabel.agent.core.Operation
 import com.embabel.agent.core.ProcessContext
+import com.embabel.agent.core.internal.LlmOperations
 import com.embabel.agent.core.support.LlmInteraction
-import com.embabel.agent.spi.LlmOperations
 import com.embabel.agent.spi.support.springai.ChatClientLlmOperations
 import com.embabel.chat.Message
 import com.embabel.common.ai.model.LlmOptions
