+++ embabel-agent-api/src/test/kotlin/com/embabel/agent/e2e/AgentPlatformIntegrationTest.kt
@@ -28,9 +28,9 @@ import com.embabel.agent.core.Agent
 import com.embabel.agent.core.AgentPlatform
 import com.embabel.agent.core.AgentProcessStatusCode
 import com.embabel.agent.core.ProcessOptions
+import com.embabel.agent.core.internal.LlmOperations
 import com.embabel.agent.domain.io.UserInput
 import com.embabel.agent.domain.library.HasContent
-import com.embabel.agent.spi.LlmOperations
 import com.embabel.agent.spi.expression.spel.Elephant
 import com.embabel.agent.spi.expression.spel.Spel2ActionsNoGoal
 import com.embabel.agent.spi.expression.spel.Zoo
